Fix RTP intercept failures if SIP is proxied back to the source IP
Bug was introduced by
https://github.com/OpenLI-NZ/openli/commit/0049127c86ec813252f02778a12e074f751dfc14

In certain setups, a SIP proxy may end up proxying a SIP session
back to the exact same IP address that sent the initial SIP
messages in the first place. In other words, we see an
INVITE from IP A -> proxy B, followed by the same (mostly)
INVITE going back from proxy B to IP A.

In that case, we were erroneously determining that the proxied
INVITE was a "reverse INVITE" and acting as though the SDP info
in that INVITE applied to the other direction of the corresponding
RTP stream. If the proxy had forwarded that INVITE anywhere else
then there wouldn't be a problem, but the destination IP matching
the source of the initial INVITE would trigger the "reverse INVITE"
logic.

This change simply forces any INVITE seen prior to the RTP stream
becoming "active", i.e. the 200 OK being seen for the most recent
INVITE, to be treated as an initial INVITE regardless of its
destination IP address.

Note to self: I guess there is still potential for things to go
haywire if a proxy is sending back to the same source IP AND a
genuine reverse INVITE occurs....