Only generate EPS CCs for GTPv2 sessions
GTPv1 sessions should be generating UMTS CCs, but that's getting
to be somewhat legacy now so we don't want to have to support that
any more than we need to.

Also tidy up some extraneous debug output and testing code that
was accidentally left behind prior to the latest merge.