Another attempt to resolve reverse INVITEs + proxying
This solution seems to work for the conflicting cases that I
had discovered so far -- we use a stack to track the details of
the RTP session for the initial INVITE sender and try to ignore
any other proxying (at least from an RTP interception perspective).

Reverse INVITEs are ignored until we see one destined for the
initial INVITE sender.

We could also now consider discarding the proxied SIP, but that's
something that will require a bit more thought.

There is still one edge case where this will fail (when the
proxied SIP is also sent back to the IP of the initial INVITE
sender AND there is a reverse INVITE) but that seems almost
impossible to resolve and hopefully won't happen any time soon.