Fix bug affecting multiple intercepts for the same RTP stream
If we had multiple intercepts configured that required the
collector to intercept the same RTP stream multiple times, not
all of those intercepts were guaranteed to take place.

The problem was due to the invitecseq string for the stream
being shared by all of those intercepts, and the string was
sometimes being freed before all of the intercepts had been
able to determine that the INVITE was successful.