Fix bug where mediator disconnections were missed by the collector
This bug occured when RMQ was being used to transfer ETSI records
between the collector and the mediator -- read events on the
regular file descriptor were ignored, so there was no way for the
collector to recognise that the peer had reset the connection.

A related issue was that subsequent connection attempts by the
mediator were discarded because the mediator IP and port matched
the one that was supposedly already connected. Therefore even if
the mediator reconnected, it would be ignored by the collector
and never receive any data unless the collector itself was
restarted.