TCP reassembly: don't replace saved packets with later retransmits
If the segment size increased on retransmit, we were replacing the
original saved packet entry with the later retransmit to try
and produce a more "correct" packet stream for RAWIP IRI outputs.

This approach has a few issues:
 * we're discarding the original transmission, but there's no
   real reason to do so since retransmissions in a raw IP capture
   shouldn't really matter. The saved packets are ONLY used for
   raw IP, not for reconstructing SIP content so we don't have
   to "get it right". Indeed, we are better off avoiding any
   changes to the packet(s) that were seen on the wire, including
   discarding those that had to be retransmitted.
 * it is a major vector for introducing random, very occasional
   segfaults due to the potential for double freeing a saved
   packet. The extremely rare nature of a TCP retransmission that
   combines two lost segments makes this type of bug very
   difficult to replicate and debug. And indeed, this type of
   crashing is occurring for some users.
 * the packet entry list is not guaranteed to be in sequence
   number order, so there is a possibility that our method of
   replacing the "most recent" entry is going to remove the wrong
   packet anyway

Instead, the packets for any segments whose size are modified in
retransmission will now be preserved (both original and
retransmit). Retransmits that are the same size as before will
continue to be discarded -- mainly because they might be
duplicates rather than retransmits and therefore it is probably
still better for us to exclude them.

Hopefully this will in turn resolve the very occasional collector
crashes that are affecting a few voice intercept users.