SIP keepalives no longer break TCP reassembly
Previously, packets containing keepalive payload were skipped
without being passed into the TCP reassembler, which meant that
the reassembler would be perpetually waiting for the keepalive
segment(s) as the first KA would contain the expected sequence
number. Now keepalives are passed into the reassembler, but there
is new logic to recognise KA patterns in the reassembled
content and move past them accordingly.

Also fixed more bugs relating to reassembling segments when a
retransmit is seen with new payload in the same segment.
 * make sure additional unseen payload is inserted into the
   segment map when a retransmit introduces new content
 * disable "fast path" check in update_tcp_reassemble_stream() if
   the content is not likely to be a single TCP packet (i.e. the
   packet content contains multiple SIP messages, or the content
   is reassembled from multiple segments.
 * also disable "fast path" for segments that contain a keep alive
   (as fast path does not include the new KA handling logic.