collector: allow user to have zero GTP, SMS, email worker threads
In particular, disabling SMS interception could be desirable for
performance reasons (i.e. to avoid having to do TCP reassembly
+ SIP parsing to determine if a packet should be handed off to an
SMS worker).

If you are doing voice intercepts, but never expect to see SMS
traffic, then it is a waste of cycles to keep constantly looking
at the SIP traffic for MESSAGE packets that are never going to
appear.