Fix memory leaks in radius module
If a RADIUS Acct Request has multiple identities (e.g. a
username and a calling station ID), we would allocate a
request object to store attributes for each identity.
However, the parsed RADIUS state for that request would only
retain the pointer to the most recent request object, leaking
any others that had been allocated earlier.

Now, if we are careful not to allocate a request object if we
already have one and the intended request ID is the same as the
ID for the existing object.