Fix memory leak when parsing tel: URIs
All From and To URIs are now strdup'd, even if osip can give
us a zero-copy pointer. While this may not be the most performant,
having some situations where a URI is malloc'd memory and others
where it is not is a potential nightmare.

Since we're often duplicating other fields from the INVITE anway,
e.g. P-Asserted-Identity, it's hard to imagine that a couple of
extra strdups are going to introduce a bottleneck that wasn't
already there.