Initial phase of adding SIP worker redirection support
STATUS: job is about halfway done -- redirect messages are sent
but still need to implement receiving and notifications.

Motivation: we have observed SIP sessions where the teardown
messages use a different 5-tuple to the standup ones (even
changing from TCP to UDP). This breaks the new multiple-sip-thread
code as we cannot guarantee that all of the messages for a SIP
session end up on the same SIP worker thread.

Idea: if a SIP worker sees a non-INVITE, non-REGISTER message for
a Call-ID that it has never seen before, it could send that
message to all of the other workers and they could check if they
have seen the INVITE for that call session. If a worker has seen
the call ID, then it can "claim" it and then future redirections
will be sent only to that worker thread.

Concerns:
 * on startup, there are going to a lot of "mid-call" messages
   being broadcast for no reason; mitigation -- maybe do not
   redirect in the first N seconds after the worker starts?
 * ZMQ hanging if we try to redirect to a worker that has been
   halted; mitigation -- set a blanket EAGAIN timeout of 1 sec
   and toss the redirect if we get EAGAIN on send
 * too much packet copying could affect performance significantly.