provisioner: add ability to write encrypted intercept config files
Will only occur if `encrypt-intercept-config-file` is set to true
in the provisioner config AND a suitable password is provided via
the -K command line argument.

Note if enabled for an unencrypted intercept config file, any
changes to the configuration made via the REST API will cause
the config file to be encrypted.

To decrypt an inadvertently encrypted config file:
  openssl enc -d -aes-256-cbc -pbkdf2 -salt -in <file> -out <outfile>
      pass <password file>

Where 'password file' is the path to the file you set as the -K
option when starting the provisioner that encrypted the file.