Start working towards adding collectors/ and mediators/ to REST API
Added timer to update the "last_seen" column for each connected
client every 5 minutes.

Added API endpoints for GET method for collectors and mediators.

Wrote code to query for all known collectors or mediators and
return them as an array of structures for the purpose of iterating
over.

Added skeleton methods to updateserver.c that will have the task
of iterating over their corresponding client array and convert
them into an array of JSON objects that we can respond with when
queried for collectors or mediators.