Allow multiple XIDs to be configured for a single intercept
In the case where a target has multiple identifiers, e.g.
a static IPv4 address and a static IPv6 subnet, or multiple
mobile phones, the X1 standard would require each identifier to
be configured using a separate X1 task.

Therefore, OpenLI needs to be able to associate multiple XIDs with
a single intercept. This is handled by adding a 'xids' parameter
to the intercept configuration accepted by provisioner, which is
a list of XIDs. Note that the old 'xid' parameter is still
supported for specifying a single XID, although it may be
converted to a list containing one item if the provisioner has
to update the intercept config file.

Another change added here is the addition of checks to prevent
users from associating the same XID with multiple intercepts. If
the user introduces duplicate XIDs by editing the config file, the
provisioner will cease running until they fix the config file. If
it happens via the REST API, then the request will be rejected and
the configuration will remain unchanged.