Payload encryption settings can now be set at the agency level
If encryption is enabled at the agency level, all intercepts for
that agency will use the encryption options for that agency
unless explicitly overridden in the intercept config (using the
same config options that previously existed).

LIID->agency mappings sent to the mediator now also include the
encryption options for each intercept (either their own
specific options, or the ones inherited from their agency). This
will allow the mediator to be able to take over the encryption
step (inside coll recv threads).

NOTE: this all requires a lot of testing to make sure that
the correct encryption state is maintained in all of the
different circumstances.