Determine collector "identifiers" from configurable parameters
Specifically, the operator ID, network element ID and the
intercept point ID (if present).

Previously, collectors were solely identified by the IP address
that they connected to the provisioner from, which worked fine
when the identifiers were primarily for internal use. Now, that
we are exposing the collectors via the REST API, we need
identifiers that are a bit more descriptive and human friendly.