Move payload encryption from collectors to mediators
The main reason for doing this is to eventually be able to
properly support integrity checks (which are calculated against
unencrypted payload).

Previously, encryption was performed by an encoding thread within
the collector as the CC or IRI was being encoded.

Now, encoder worker will place the payload inside the same
"encryptedContainer" structure, but the encryption type will be
set to None and the unencrypted payload will be copied into the
record. Note that the payload is padded to ensure that the
container is the correct size for the eventual encrypted payload.

When the mediator receives the record, it now knows that
encryption is required (thanks to addition of encryption parameters
to the LIID->agency mapping structure), and will encrypt the
payload in-place and set the encryption type to the correct value.

Initial test results look good, but more will be required before
we can finalize this feature.