collector: forwarder will retry RMQ connection if first attempt fails
Previously, if the local rabbitmq server was down then the
forwarding thread would immediately shutdown and then get stuck
in a 100% CPU loop.

Now, the forwarded will try every second until it connects to
rabbitmq, which at least gives the user a change to realise the
problem (or for RMQ to start up if systemd has started the
collector first).