Fix memory error when encrypting in place
By default, EVP always adds padding when encrypting. This was
causing memory errors, as we had already provided padding to
satisfy the requirements in Annex G of ETSI TS 102 232-1, and
there was no space in the buffer to fit the padding added by EVP.

Instead, we can tell EVP not to add padding and that solves
that problem.