Add configure check for rabbitmq-c header location
Newer versions of rabbitmq-c put the headers in a rabbitmq-c/
subdirectory and have marked the old locations as deprecated.

However, we still need to support older versions that pre-date
this change, so we need to add a ./configure condition to
determine which version we are working with.

In the process, I've tidied up some unnecessary RMQ header
includes in various source files.