export buffer: add "retransmit window" capability
In this case, the "retransmit window" is the number of bytes
that we have previously sent to the client that we should retain
and retransmit if the client handover disconnects and reconnects.

Previously, we had a retransmit window that was always zero, i.e.
if we sent it, then we assumed the client received it. In certain
situations this may not be the case. Therefore, some agencies may
require that we retransmit a certain amount of data upon a
handover reconnection to ensure that they do not miss data that
we had sent but they didn't receive due to the connection
malfunction.