Ensure all ICS signing requests are sent at the right time
The sign timer was not being set correctly for signatures
after the first one.

Also made a fix to ensure that the sign timer is always halted
when a sign request is sent (previously, we were only halting
the timer if it triggered, which missed cases where the request
was sent due to the hash PDU count being exceeded).