Add support for LIIDs that are not pure ASCII
LIIDs in .nl are expressed using the hex representation of
25 binary octets, and thus many of these octets will fall outside
the range of values that correspond to "printable" characters.

Our approach: continue to internally treat LIIDs as strings, but
tag them to indicate whether the string needs to be converted from
a hex representation to binary octets at encoding time.

A binary LIID is configured using the hex representation of the
octets, preceded by '0x' (much like the binary encryption keys).
All other LIID formats are assumed to be ASCII strings and will
be encoded as such.

Internally, we strip the '0x' and use the hex string to act as the
LIID for intercept identification purposes. This means a lot of
code that uses the LIID (e.g. as a hash map key, or simply for
logging purposes) does not need to change.

I've also updated the encoding functions for message types where
the LIID is included in the message content (e.g. EPSCCs and
HI1 notifications) to ensure that a hex string LIID will be
encoded properly.