Add basic outline of a UDP sink worker thread to collector
sync thread keeps track of the set of valid UDP sinks for the
collector, then spins up the thread if it receives an IP intercept
which requires a sink that is in its configuration.

Plenty still to implement but this is the basic framework of how
the worker threads will be managed.