REST API: don't enforce encryption key format if encryption is disabled
Submitting an empty string as the encryption key shouldn't matter
if encryption is not enabled for an intercept anyway.

This fixes issues where the web-provisioner (which does currently
include a "" encryptkey in the API request if it is not set in
the web form) would fail to submit an intercept where the key was
blank when the method was set to 'none'.

Side effect: if the user does provide a valid key in their
request but sets the method to 'none', the key will not be saved.
Similarly, if they modify an existing encrypted intercept to
disable encryption, the old key will be lost.