Add code to generate IPIRIs when a sink begins/ends interception
Note: at the moment, each sink will generate IPIRIs independently.
This means that if multiple sinks are responsible for the same
intercept (e.g. one sink for each direction), then we are going
to produce duplicate IPIRIs.

The next step is to add some de-duplication in the sequence
tracker threads, as this is most likely the simplest place to
handle this situation.