Add support for Nokia mirroring to UDP sinks
Refactored ALUshim parsing code to support being called
in a context where we assume the payload is intended to be
intercepted (as in a UDP sink).

check_alu_intercept() still does the other checks that are
required if the mirrored traffic is mixed in with other traffic
sources, whereas decode_alushim_from_udp_payload() just parses
the shim and finds the IP header in the encapsulated content.