Action changes to UDP sink configuration in collector config
If the UDP sinks defined in the collector config are changed,
then the collector sync thread will now start/halt UDP sink
threads accordingly.

One tricky aspect of this change is that we can now no longer
ignore announcements that map intercepts to UDP sinks that a
collector currently isn't running, as it is now possible that
a user may add the sink to the collector configuration after
configuring the intercept itself (or have to correct an error
in the existing collector config).

Instead, we have to keep a map of "unused" UDP sink intercept
configurations so we can apply the right arguments to the sink
thread if we need to start it later on.

One thing that we don't protect against though is the removal of
a UDP sink that was currently being used for an active intercept.
But I'm not sure that a) this would not get in the way of
legitimate behaviour (e.g. fixing configuration errors), and b)
that there is a good way to implement it anyway since the user
will have already modified and saved their config file to disk.