Fix missing IRIs for re-added static IP intercepts that re-use session IDs
Technically, the session ID should change whenever a static IP
intercept is removed and then re-added but, in practice, that
is probably not something that a user will consider if they
are re-adding an intercept.

Instead, we now check for the appearance of a BEGIN IRI after
the intercept has been removed that matches the session ID of
the previous intercept. If we see that, we assume the user has
re-added the intercept without changing the session ID and
simply reset the reordering state to allow the BEGIN IRI to
be passed on to the mediator.

It's not ideal and may be frustrating for the receiving LEA to
deal with, but I think it is a more intuitive behaviour than
not producing the BEGIN IRI at all.