SIP: replace invitecseq_stack with a simpler method
The existing method falls over when SIP INVITEs end up being
retransmitted in a way that does not result in an additional
response, e.g. if the INVITE gets lost and has to be re-sent.

In that case, the stack count grows but there are not enough
responses to get it back to zero, so the other side of the RTP
stream is never saved.

The simpler method is to just assume any response with the exact
same CSeq, port and address (destination) as the initial INVITE
is a suitable response for us to use for RTP tracking purposes.