Fix bug where REST API would return a "-1" session ID for UDP sinks
This would occur if the session ID has not been configured for
the sink. Now if the session ID is not set, we will not include
a "sessionid" field in the JSON object for the sink.

If the session ID is not set and the sink is using an
encapsulation that does not include one, we will continue to
use "1" as the CIN for anything intercepted by that sink.