Minor optimization for processing UDP traffic
Don't waste time calling trace_get_timeval() on packets if there
are no current RTP streams that need to be intercepted.

Obviously, this only helps in cases where no interceptable calls
are active -- so it is not really much of a gain in practice, as
once you have to intercept a call then you are back to calling
trace_get_timeval() on every packet.